requirements.txt
Flask
openai
faiss-cpu
numpy
requests
nltk
tqdm
gunicorn
markdown2